requirements.txt
requests
pyotp
python-dotenv
kiteconnect
pandas
PyYAML
rich
beepy
dhanhq